changed int to byte

+++ graphics-common/src/main/java/org/sudu/experiments/math/ArrayOp.java
@@ -108,7 +108,7 @@ public interface ArrayOp {
     return array.length == size ? array : Arrays.copyOf(array, size);
   }
 
-  static int[] resizeOrReturn(int[] array, int size) {
+  static byte[] resizeOrReturn(byte[] array, int size) {
     return array.length == size ? array : Arrays.copyOf(array, size);
   }
 
@@ -120,7 +120,7 @@ public interface ArrayOp {
     return data;
   }
 
-  static int[] addAt(int value, int[] data, int index) {
+  static byte[] addAt(byte value, byte[] data, int index) {
     if (data.length == index) {
       data = Arrays.copyOf(data, data.length * 2);
     }
